requirements.txt
python-dotenv
requests
Pillow
aiohttp
azure-sdk-tools
azure-core
azure-ai-vision-imageanalysis